Move buildCharmCard to own file
